Refine cost summary display settings
