+++ Sources/CodexBar/Resources/ar.lproj/Localizable.strings
@@ -418,9 +418,14 @@
 "show_cost_summary" = "ملخص تكلفة العرض";
 "show_cost_summary_subtitle" = "يقرأ سجلات الاستخدام المحلية. يعرض اليوم + نافذة التاريخ المحددة في القائمة.";
 "cost_summary_style_title" = "نمط العرض";
-"cost_summary_style_inline" = "ملخص مضمن";
-"cost_summary_style_submenu" = "قائمة التكلفة الفرعية";
+"cost_summary_style_inline" = "مضمن فقط";
+"cost_summary_style_submenu" = "قائمة فرعية فقط";
 "cost_summary_style_both" = "كلاهما";
+"cost_summary_style_inline_help" = "يعرض ملخص التكلفة مباشرة في القائمة الرئيسية.";
+"cost_summary_style_submenu_help" = "يعرض بدلاً من ذلك قائمة التكلفة الفرعية التفصيلية.";
+"cost_summary_style_both_help" = "يعرض ملخص القائمة الرئيسية وقائمة التكلفة الفرعية التفصيلية.";
+"cost_history_window_title" = "نافذة التاريخ";
+"cost_history_window_help" = "يحدد عدد أيام سجلات الاستخدام المحلية التي تظهر في القائمة.";
 "cost_history_days_title" = "نافذة التاريخ: %d أيام";
 "cost_auto_refresh_info" = "تحديث تلقائي: كل ساعة · وقت الاستراحة: 10m";
 "refresh_cadence_title" = "وتيرة التحديث";
