+++ Sources/CodexBar/Resources/ca.lproj/Localizable.strings
@@ -416,9 +416,14 @@
 "show_cost_summary" = "Mostreu el resum de cost";
 "show_cost_summary_subtitle" = "Llegeix els registres d'ús locals. Mostra el cost d'avui + la finestra d'historial seleccionada al menú.";
 "cost_summary_style_title" = "Estil de visualització";
-"cost_summary_style_inline" = "Resum integrat";
-"cost_summary_style_submenu" = "Submenú de cost";
+"cost_summary_style_inline" = "Només integrat";
+"cost_summary_style_submenu" = "Només submenú";
 "cost_summary_style_both" = "Tots dos";
+"cost_summary_style_inline_help" = "Mostra el resum de cost directament al menú principal.";
+"cost_summary_style_submenu_help" = "Mostra el submenú Cost detallat en lloc d'això.";
+"cost_summary_style_both_help" = "Mostra el resum del menú principal i el submenú Cost detallat.";
+"cost_history_window_title" = "Finestra d'historial";
+"cost_history_window_help" = "Defineix quants dies de registres d'ús locals apareixen al menú.";
 "cost_history_days_title" = "Finestra d'historial: %d dies";
 "cost_auto_refresh_info" = "Actualització automàtica: cada hora · Temps d'espera: 10 min";
 "refresh_cadence_title" = "Freqüència d'actualització";
