+++ Sources/CodexBar/Resources/de.lproj/Localizable.strings
@@ -418,9 +418,14 @@
 "show_cost_summary" = "Kostenübersicht anzeigen";
 "show_cost_summary_subtitle" = "Liest lokale Nutzungsprotokolle. Zeigt heute + das gewählte Verlaufsfenster im Menü.";
 "cost_summary_style_title" = "Anzeigestil";
-"cost_summary_style_inline" = "Inline-Übersicht";
-"cost_summary_style_submenu" = "Kosten-Untermenü";
+"cost_summary_style_inline" = "Nur Inline";
+"cost_summary_style_submenu" = "Nur Untermenü";
 "cost_summary_style_both" = "Beides";
+"cost_summary_style_inline_help" = "Zeigt die Kostenübersicht direkt im Hauptmenü.";
+"cost_summary_style_submenu_help" = "Zeigt stattdessen das detaillierte Kosten-Untermenü.";
+"cost_summary_style_both_help" = "Zeigt die Hauptmenü-Übersicht und das detaillierte Kosten-Untermenü.";
+"cost_history_window_title" = "Verlaufsfenster";
+"cost_history_window_help" = "Legt fest, wie viele Tage lokaler Nutzungsprotokolle im Menü erscheinen.";
 "cost_history_days_title" = "Verlaufsfenster: %d Tage";
 "cost_auto_refresh_info" = "Auto-Aktualisierung: stündlich · Timeout: 10 Min.";
 "refresh_cadence_title" = "Aktualisierungsintervall";
