+++ Sources/CodexBar/Resources/en.lproj/Localizable.strings
@@ -418,9 +418,14 @@
 "show_cost_summary" = "Show cost summary";
 "show_cost_summary_subtitle" = "Reads local usage logs. Shows today + the selected history window in the menu.";
 "cost_summary_style_title" = "Display style";
-"cost_summary_style_inline" = "Inline";
-"cost_summary_style_submenu" = "Submenu";
+"cost_summary_style_inline" = "Inline only";
+"cost_summary_style_submenu" = "Submenu only";
 "cost_summary_style_both" = "Both";
+"cost_summary_style_inline_help" = "Shows the cost summary directly in the main menu.";
+"cost_summary_style_submenu_help" = "Shows the detailed Cost submenu instead.";
+"cost_summary_style_both_help" = "Shows both the main menu summary and detailed Cost submenu.";
+"cost_history_window_title" = "History window";
+"cost_history_window_help" = "Sets how many days of local usage logs appear in the menu.";
 "cost_history_days_title" = "History window: %d days";
 "cost_auto_refresh_info" = "Auto-refresh: hourly · Timeout: 10m";
 "refresh_cadence_title" = "Refresh cadence";
