+++ Sources/CodexBar/Resources/es.lproj/Localizable.strings
@@ -416,9 +416,14 @@
 "show_cost_summary" = "Mostrar resumen de coste";
 "show_cost_summary_subtitle" = "Lee los registros de uso locales. Muestra el coste de hoy + la ventana de historial seleccionada en el menú.";
 "cost_summary_style_title" = "Estilo de visualización";
-"cost_summary_style_inline" = "Resumen integrado";
-"cost_summary_style_submenu" = "Submenú de coste";
+"cost_summary_style_inline" = "Solo integrado";
+"cost_summary_style_submenu" = "Solo submenú";
 "cost_summary_style_both" = "Ambos";
+"cost_summary_style_inline_help" = "Muestra el resumen de coste directamente en el menú principal.";
+"cost_summary_style_submenu_help" = "Muestra en su lugar el submenú Coste detallado.";
+"cost_summary_style_both_help" = "Muestra el resumen del menú principal y el submenú Coste detallado.";
+"cost_history_window_title" = "Ventana de historial";
+"cost_history_window_help" = "Define cuántos días de registros de uso locales aparecen en el menú.";
 "cost_history_days_title" = "Ventana de historial: %d días";
 "cost_auto_refresh_info" = "Actualización automática: cada hora · Tiempo de espera: 10 m";
 "refresh_cadence_title" = "Frecuencia de actualización";
