+++ Sources/CodexBar/Resources/fa.lproj/Localizable.strings
@@ -418,9 +418,14 @@
 "show_cost_summary" = "خلاصه هزینه نمایش";
 "show_cost_summary_subtitle" = "گزارش های استفاده محلی را می خواند. امروز نمایش داده می شود + پنجره تاریخچه انتخاب شده در منو.";
 "cost_summary_style_title" = "سبک نمایش";
-"cost_summary_style_inline" = "خلاصه درون‌خطی";
-"cost_summary_style_submenu" = "زیرمنوی هزینه";
+"cost_summary_style_inline" = "فقط درون‌خطی";
+"cost_summary_style_submenu" = "فقط زیرمنو";
 "cost_summary_style_both" = "هر دو";
+"cost_summary_style_inline_help" = "خلاصه هزینه را مستقیماً در منوی اصلی نشان می‌دهد.";
+"cost_summary_style_submenu_help" = "در عوض زیرمنوی جزئیات هزینه را نشان می‌دهد.";
+"cost_summary_style_both_help" = "هم خلاصه منوی اصلی و هم زیرمنوی جزئیات هزینه را نشان می‌دهد.";
+"cost_history_window_title" = "پنجره تاریخچه";
+"cost_history_window_help" = "تعیین می‌کند چند روز از گزارش‌های استفاده محلی در منو نشان داده شود.";
 "cost_history_days_title" = "پنجره تاریخچه: %d روز";
 "cost_auto_refresh_info" = "تازه سازی خودکار: ساعتی · زمان استراحت: 10m";
 "refresh_cadence_title" = "کادانس تازه سازی";
