+++ Sources/CodexBar/Resources/fr.lproj/Localizable.strings
@@ -418,9 +418,14 @@
 "show_cost_summary" = "Afficher le récapitulatif des coûts";
 "show_cost_summary_subtitle" = "Lit les journaux d'utilisation locaux. Affiche le coût d'aujourd'hui et de la période sélectionnée dans le menu.";
 "cost_summary_style_title" = "Style d’affichage";
-"cost_summary_style_inline" = "Résumé intégré";
-"cost_summary_style_submenu" = "Sous-menu des coûts";
+"cost_summary_style_inline" = "Intégré seulement";
+"cost_summary_style_submenu" = "Sous-menu seulement";
 "cost_summary_style_both" = "Les deux";
+"cost_summary_style_inline_help" = "Affiche le résumé des coûts directement dans le menu principal.";
+"cost_summary_style_submenu_help" = "Affiche plutôt le sous-menu Coût détaillé.";
+"cost_summary_style_both_help" = "Affiche le résumé du menu principal et le sous-menu Coût détaillé.";
+"cost_history_window_title" = "Fenêtre d'historique";
+"cost_history_window_help" = "Définit le nombre de jours de journaux d'utilisation locaux affichés dans le menu.";
 "cost_history_days_title" = "Fenêtre d'historique : %d jours";
 "cost_auto_refresh_info" = "Actualisation automatique : toutes les heures · Délai d'expiration : 10 min";
 "refresh_cadence_title" = "Fréquence d'actualisation";
