+++ Sources/CodexBar/Resources/id.lproj/Localizable.strings
@@ -420,9 +420,14 @@
 "show_cost_summary" = "Tampilkan ringkasan biaya";
 "show_cost_summary_subtitle" = "Membaca log penggunaan lokal. Menampilkan hari ini + jendela riwayat yang dipilih di menu.";
 "cost_summary_style_title" = "Gaya tampilan";
-"cost_summary_style_inline" = "Ringkasan sebaris";
-"cost_summary_style_submenu" = "Submenu biaya";
+"cost_summary_style_inline" = "Hanya sebaris";
+"cost_summary_style_submenu" = "Hanya submenu";
 "cost_summary_style_both" = "Keduanya";
+"cost_summary_style_inline_help" = "Menampilkan ringkasan biaya langsung di menu utama.";
+"cost_summary_style_submenu_help" = "Menampilkan submenu Biaya terperinci sebagai gantinya.";
+"cost_summary_style_both_help" = "Menampilkan ringkasan menu utama dan submenu Biaya terperinci.";
+"cost_history_window_title" = "Jendela riwayat";
+"cost_history_window_help" = "Menentukan berapa hari log penggunaan lokal yang ditampilkan di menu.";
 "cost_history_days_title" = "Jendela riwayat: %d hari";
 "cost_auto_refresh_info" = "Penyegaran otomatis: per jam · Batas waktu: 10m";
 "refresh_cadence_title" = "Frekuensi penyegaran";
