+++ Sources/CodexBar/Resources/it.lproj/Localizable.strings
@@ -420,9 +420,14 @@
 "show_cost_summary" = "Mostra riepilogo costi";
 "show_cost_summary_subtitle" = "Legge i log di utilizzo locali. Mostra oggi + la finestra storica selezionata nel menu.";
 "cost_summary_style_title" = "Stile di visualizzazione";
-"cost_summary_style_inline" = "Riepilogo in linea";
-"cost_summary_style_submenu" = "Sottomenu costi";
+"cost_summary_style_inline" = "Solo in linea";
+"cost_summary_style_submenu" = "Solo sottomenu";
 "cost_summary_style_both" = "Entrambi";
+"cost_summary_style_inline_help" = "Mostra il riepilogo dei costi direttamente nel menu principale.";
+"cost_summary_style_submenu_help" = "Mostra invece il sottomenu Costo dettagliato.";
+"cost_summary_style_both_help" = "Mostra sia il riepilogo nel menu principale sia il sottomenu Costo dettagliato.";
+"cost_history_window_title" = "Finestra storica";
+"cost_history_window_help" = "Imposta quanti giorni di log di utilizzo locali mostrare nel menu.";
 "cost_history_days_title" = "Finestra storica: %d giorni";
 "cost_auto_refresh_info" = "Aggiornamento automatico: ogni ora · Timeout: 10 min";
 "refresh_cadence_title" = "Frequenza aggiornamento";
