+++ Sources/CodexBar/Resources/ja.lproj/Localizable.strings
@@ -415,9 +415,14 @@
 "show_cost_summary" = "コスト概要を表示";
 "show_cost_summary_subtitle" = "ローカルの使用ログを読み取り、今日分と選択した履歴期間をメニューに表示します。";
 "cost_summary_style_title" = "表示スタイル";
-"cost_summary_style_inline" = "インライン概要";
-"cost_summary_style_submenu" = "コストサブメニュー";
+"cost_summary_style_inline" = "インラインのみ";
+"cost_summary_style_submenu" = "サブメニューのみ";
 "cost_summary_style_both" = "両方";
+"cost_summary_style_inline_help" = "コスト概要をメインメニューに直接表示します。";
+"cost_summary_style_submenu_help" = "代わりに詳細なコストサブメニューを表示します。";
+"cost_summary_style_both_help" = "メインメニューの概要と詳細なコストサブメニューの両方を表示します。";
+"cost_history_window_title" = "履歴期間";
+"cost_history_window_help" = "メニューに表示するローカル使用ログの日数を設定します。";
 "cost_history_days_title" = "履歴期間: %d 日";
 "cost_auto_refresh_info" = "自動更新: 1 時間ごと · タイムアウト: 10 分";
 "refresh_cadence_title" = "更新間隔";
