+++ Sources/CodexBar/Resources/nl.lproj/Localizable.strings
@@ -418,9 +418,14 @@
 "show_cost_summary" = "Kostenoverzicht weergeven";
 "show_cost_summary_subtitle" = "Leest lokale gebruikslogboeken. Toont vandaag + het geselecteerde geschiedenisvenster in het menu.";
 "cost_summary_style_title" = "Weergavestijl";
-"cost_summary_style_inline" = "Inline-overzicht";
-"cost_summary_style_submenu" = "Kosten-submenu";
+"cost_summary_style_inline" = "Alleen inline";
+"cost_summary_style_submenu" = "Alleen submenu";
 "cost_summary_style_both" = "Beide";
+"cost_summary_style_inline_help" = "Toont het kostenoverzicht rechtstreeks in het hoofdmenu.";
+"cost_summary_style_submenu_help" = "Toont in plaats daarvan het gedetailleerde Kosten-submenu.";
+"cost_summary_style_both_help" = "Toont zowel het hoofdmenu-overzicht als het gedetailleerde Kosten-submenu.";
+"cost_history_window_title" = "Geschiedenisvenster";
+"cost_history_window_help" = "Stelt in hoeveel dagen lokale gebruikslogboeken in het menu verschijnen.";
 "cost_history_days_title" = "Geschiedenisvenster: %d dagen";
 "cost_auto_refresh_info" = "Automatisch vernieuwen: elk uur · Time-out: 10m";
 "refresh_cadence_title" = "Cadans vernieuwen";
