+++ Sources/CodexBar/Resources/pl.lproj/Localizable.strings
@@ -420,9 +420,14 @@
 "show_cost_summary" = "Pokaż podsumowanie kosztów";
 "show_cost_summary_subtitle" = "Odczytuje lokalne logi użycia. Pokazuje dziś + wybrany zakres historii w menu.";
 "cost_summary_style_title" = "Styl wyświetlania";
-"cost_summary_style_inline" = "Wbudowane podsumowanie";
-"cost_summary_style_submenu" = "Podmenu kosztów";
+"cost_summary_style_inline" = "Tylko wbudowane";
+"cost_summary_style_submenu" = "Tylko podmenu";
 "cost_summary_style_both" = "Oba";
+"cost_summary_style_inline_help" = "Pokazuje podsumowanie kosztów bezpośrednio w menu głównym.";
+"cost_summary_style_submenu_help" = "Zamiast tego pokazuje szczegółowe podmenu Koszt.";
+"cost_summary_style_both_help" = "Pokazuje podsumowanie w menu głównym i szczegółowe podmenu Koszt.";
+"cost_history_window_title" = "Zakres historii";
+"cost_history_window_help" = "Ustawia, ile dni lokalnych dzienników użycia pokazać w menu.";
 "cost_history_days_title" = "Zakres historii: %d dni";
 "cost_auto_refresh_info" = "Auto-odświeżanie: co godzinę · Limit czasu: 10 min";
 "refresh_cadence_title" = "Częstotliwość odświeżania";
