+++ Sources/CodexBar/Resources/pt-BR.lproj/Localizable.strings
@@ -415,9 +415,14 @@
 "show_cost_summary" = "Mostrar resumo de custos";
 "show_cost_summary_subtitle" = "Lê logs de uso locais. Mostra o custo de hoje + janela selecionada no menu.";
 "cost_summary_style_title" = "Estilo de exibição";
-"cost_summary_style_inline" = "Resumo embutido";
-"cost_summary_style_submenu" = "Submenu de custos";
+"cost_summary_style_inline" = "Somente embutido";
+"cost_summary_style_submenu" = "Somente submenu";
 "cost_summary_style_both" = "Ambos";
+"cost_summary_style_inline_help" = "Mostra o resumo de custos diretamente no menu principal.";
+"cost_summary_style_submenu_help" = "Mostra o submenu Custo detalhado em vez disso.";
+"cost_summary_style_both_help" = "Mostra o resumo no menu principal e o submenu Custo detalhado.";
+"cost_history_window_title" = "Janela do histórico";
+"cost_history_window_help" = "Define quantos dias de logs de uso locais aparecem no menu.";
 "cost_history_days_title" = "Janela do histórico: %d dias";
 "cost_auto_refresh_info" = "Atualização automática: a cada hora · Timeout: 10 min";
 "refresh_cadence_title" = "Cadência de atualização";
