+++ Sources/CodexBar/Resources/sv.lproj/Localizable.strings
@@ -418,9 +418,14 @@
 "show_cost_summary" = "Visa kostnadssammanfattning";
 "show_cost_summary_subtitle" = "Läser lokala användningsloggar. Visar idag och valt historikfönster i menyn.";
 "cost_summary_style_title" = "Visningsstil";
-"cost_summary_style_inline" = "Inbäddad sammanfattning";
-"cost_summary_style_submenu" = "Kostnadsundermeny";
+"cost_summary_style_inline" = "Endast inbäddad";
+"cost_summary_style_submenu" = "Endast undermeny";
 "cost_summary_style_both" = "Båda";
+"cost_summary_style_inline_help" = "Visar kostnadssammanfattningen direkt i huvudmenyn.";
+"cost_summary_style_submenu_help" = "Visar den detaljerade Kostnad-undermenyn i stället.";
+"cost_summary_style_both_help" = "Visar både huvudmenyns sammanfattning och den detaljerade Kostnad-undermenyn.";
+"cost_history_window_title" = "Historikfönster";
+"cost_history_window_help" = "Anger hur många dagar med lokala användningsloggar som visas i menyn.";
 "cost_history_days_title" = "Historikfönster: %d dagar";
 "cost_auto_refresh_info" = "Automatisk uppdatering: varje timme · Timeout: 10 min";
 "refresh_cadence_title" = "Uppdateringsintervall";
