+++ Sources/CodexBar/Resources/th.lproj/Localizable.strings
@@ -418,9 +418,14 @@
 "show_cost_summary" = "แสดงสรุปค่าใช้จ่าย";
 "show_cost_summary_subtitle" = "อ่านบันทึกการใช้งานในเครื่อง แสดงวันนี้ + หน้าต่างประวัติที่เลือกในเมนู";
 "cost_summary_style_title" = "รูปแบบการแสดงผล";
-"cost_summary_style_inline" = "สรุปแบบอินไลน์";
-"cost_summary_style_submenu" = "เมนูย่อยค่าใช้จ่าย";
+"cost_summary_style_inline" = "อินไลน์เท่านั้น";
+"cost_summary_style_submenu" = "เมนูย่อยเท่านั้น";
 "cost_summary_style_both" = "ทั้งสอง";
+"cost_summary_style_inline_help" = "แสดงสรุปค่าใช้จ่ายในเมนูหลักโดยตรง";
+"cost_summary_style_submenu_help" = "แสดงเมนูย่อยค่าใช้จ่ายแบบละเอียดแทน";
+"cost_summary_style_both_help" = "แสดงทั้งสรุปในเมนูหลักและเมนูย่อยค่าใช้จ่ายแบบละเอียด";
+"cost_history_window_title" = "กรอบเวลาประวัติ";
+"cost_history_window_help" = "กำหนดจำนวนวันของบันทึกการใช้งานในเครื่องที่จะแสดงในเมนู";
 "cost_history_days_title" = "กรอบเวลาประวัติ: %d วัน";
 "cost_auto_refresh_info" = "รีเฟรชอัตโนมัติ: รายชั่วโมง · หมดเวลา: 10m";
 "refresh_cadence_title" = "จังหวะการรีเฟรช";
