+++ Sources/CodexBar/Resources/tr.lproj/Localizable.strings
@@ -418,9 +418,14 @@
 "show_cost_summary" = "Maliyet özetini göster";
 "show_cost_summary_subtitle" = "Yerel kullanım günlüklerini okur. Menüde bugün + seçilen geçmiş penceresini gösterir.";
 "cost_summary_style_title" = "Görüntüleme stili";
-"cost_summary_style_inline" = "Satır içi özet";
-"cost_summary_style_submenu" = "Maliyet alt menüsü";
+"cost_summary_style_inline" = "Yalnızca satır içi";
+"cost_summary_style_submenu" = "Yalnızca alt menü";
 "cost_summary_style_both" = "İkisi de";
+"cost_summary_style_inline_help" = "Maliyet özetini doğrudan ana menüde gösterir.";
+"cost_summary_style_submenu_help" = "Bunun yerine ayrıntılı Maliyet alt menüsünü gösterir.";
+"cost_summary_style_both_help" = "Hem ana menü özetini hem de ayrıntılı Maliyet alt menüsünü gösterir.";
+"cost_history_window_title" = "Geçmiş penceresi";
+"cost_history_window_help" = "Menüde kaç günlük yerel kullanım günlüğünün gösterileceğini belirler.";
 "cost_history_days_title" = "Geçmiş penceresi: %d gün";
 "cost_auto_refresh_info" = "Otomatik yenileme: saatlik · Zaman aşımı: 10 dk";
 "refresh_cadence_title" = "Yenileme sıklığı";
