+++ Sources/CodexBar/Resources/uk.lproj/Localizable.strings
@@ -418,9 +418,14 @@
 "show_cost_summary" = "Показати підсумок витрат";
 "show_cost_summary_subtitle" = "Читає локальні журнали використання. Показує сьогодні + вибране вікно історії в меню.";
 "cost_summary_style_title" = "Стиль відображення";
-"cost_summary_style_inline" = "Вбудований підсумок";
-"cost_summary_style_submenu" = "Підменю витрат";
+"cost_summary_style_inline" = "Лише вбудовано";
+"cost_summary_style_submenu" = "Лише підменю";
 "cost_summary_style_both" = "Обидва";
+"cost_summary_style_inline_help" = "Показує підсумок витрат безпосередньо в головному меню.";
+"cost_summary_style_submenu_help" = "Натомість показує детальне підменю Вартість.";
+"cost_summary_style_both_help" = "Показує підсумок у головному меню та детальне підменю Вартість.";
+"cost_history_window_title" = "Вікно історії";
+"cost_history_window_help" = "Визначає, скільки днів локальних журналів використання показувати в меню.";
 "cost_history_days_title" = "Вікно історії: %d днів";
 "cost_auto_refresh_info" = "Автоматичне оновлення: щогодини · Час очікування: 10 хв";
 "refresh_cadence_title" = "Оновити каденцію";
