+++ Sources/CodexBar/Resources/vi.lproj/Localizable.strings
@@ -414,9 +414,14 @@
 "show_cost_summary" = "Hiển thị tóm tắt chi phí";
 "show_cost_summary_subtitle" = "Đọc nhật ký Mức sử dụng cục bộ. Hiển thị hôm nay + cửa sổ lịch sử đã chọn trong menu.";
 "cost_summary_style_title" = "Kiểu hiển thị";
-"cost_summary_style_inline" = "Tóm tắt nội tuyến";
-"cost_summary_style_submenu" = "Menu con chi phí";
+"cost_summary_style_inline" = "Chỉ nội tuyến";
+"cost_summary_style_submenu" = "Chỉ menu con";
 "cost_summary_style_both" = "Cả hai";
+"cost_summary_style_inline_help" = "Hiển thị tóm tắt chi phí trực tiếp trong menu chính.";
+"cost_summary_style_submenu_help" = "Thay vào đó hiển thị menu con Chi phí chi tiết.";
+"cost_summary_style_both_help" = "Hiển thị cả tóm tắt menu chính và menu con Chi phí chi tiết.";
+"cost_history_window_title" = "Cửa sổ lịch sử";
+"cost_history_window_help" = "Đặt số ngày nhật ký sử dụng cục bộ xuất hiện trong menu.";
 "cost_history_days_title" = "Cửa sổ lịch sử: %d ngày";
 "cost_auto_refresh_info" = "Tự động làm mới: hàng giờ · Thời gian chờ: 10 phút";
 "refresh_cadence_title" = "Nhịp làm mới";
