+++ Sources/CodexBar/Resources/zh-Hant.lproj/Localizable.strings
@@ -422,9 +422,14 @@
 "show_cost_summary" = "顯示費用摘要";
 "show_cost_summary_subtitle" = "讀取本機使用量記錄。在選單中顯示今天及所選歷史時段的費用。";
 "cost_summary_style_title" = "顯示樣式";
-"cost_summary_style_inline" = "內嵌摘要";
-"cost_summary_style_submenu" = "費用子選單";
+"cost_summary_style_inline" = "僅內嵌";
+"cost_summary_style_submenu" = "僅子選單";
 "cost_summary_style_both" = "兩者";
+"cost_summary_style_inline_help" = "直接在主選單中顯示費用摘要。";
+"cost_summary_style_submenu_help" = "改為顯示詳細的費用子選單。";
+"cost_summary_style_both_help" = "同時顯示主選單摘要和詳細的費用子選單。";
+"cost_history_window_title" = "歷史時段";
+"cost_history_window_help" = "設定選單中顯示多少天的本機使用記錄。";
 "cost_history_days_title" = "歷史時段：%d 天";
 "cost_auto_refresh_info" = "自動重新整理：每小時 · 逾時：10 分鐘";
 "refresh_cadence_title" = "重新整理頻率";
